+++ src/util/Types.tsx
@@ -77,51 +77,3 @@ export type AttendanceRecord = {
 };
 
 export type AttendanceTag = "K" | "A" | "L" | "D" | "O" | "N" | "E";
-
-export const AttendanceTagToText: Record<AttendanceTag, string> = {
-  K: "Absent",
-  A: "Absent",
-  L: "Late Arrival",
-  D: "Early Dismissal",
-  O: "Attendend",
-  N: "Not Required ",
-  E: "Excused ",
-};
-
-export const AttendanceTagToClassName: Record<AttendanceTag, string> = {
-  L: "bg-attendance-yellow border border-attendance-yellow bg-opacity-25",
-  D: "bg-attendance-yellow border border-attendance-yellow bg-opacity-25",
-  O: "bg-attendance-green border border-attendance-green bg-opacity-25",
-  A: "bg-attendance-red border border-attendance-red bg-opacity-25",
-  K: "bg-attendance-red border border-attendance-red bg-opacity-25",
-  E: "bg-attendance-grey border border-attendance-grey bg-opacity-25",
-  N: "bg-attendance-grey border border-attendance-grey bg-opacity-25",
-};
-
-export type AttendanceStandingDisplay = {
-  text: string;
-  className: string;
-};
-
-export const AttendanceStandingOptions: AttendanceStandingDisplay[] = [
-  {
-    text: "Removed",
-    className:
-      "bg-attendance-grey border border-attendance-grey bg-opacity-25 text-atn-disabled-text",
-  },
-  {
-    text: "On hold",
-    className:
-      "bg-attendance-yellow border border-attendance-yellow bg-opacity-25 text-attendance-text-yellow",
-  },
-  {
-    text: "At risk",
-    className:
-      "bg-attendance-red border border-attendance-red bg-opacity-25 text-red-500",
-  },
-  {
-    text: "Good standing",
-    className:
-      "bg-attendance-green border border-attendance-green bg-opacity-25 text-attendance-text-green",
-  },
-];
